+++ lib/controllers/addOn.js
@@ -1,7 +1,7 @@
 var librarian = require('../common/api').librarian
 var userConfig = require('../common/api').userConfig
+var apiUtil = require('../common/api').apiUtil
 var Errors = require('../util/errors')
-var apiUtil = require('@xervo/api-util')
 
 // -----------------------------------------------------------------------------
 var AddOn = function () {
@@ -32,16 +32,7 @@ AddOn.prototype.getForProject = function (projectId, callback) {
 
 // -----------------------------------------------------------------------------
 AddOn.prototype.getConfigVars = function (projectId, callback) {
-  var api = userConfig.data.api_uri.split('://')
-
-  var Client = apiUtil({
-    protocol: api[0],
-    hostname: api[1],
-    version: 'v2',
-    token: userConfig.data.apiKey
-  })
-
-  Client.get('/projects/' + projectId + '/environment-variables', function (err, vars) {
+  apiUtil.get('/projects/' + projectId + '/environment-variables', function (err, vars) {
     if (err) {
       return callback(Errors.getMessage(err))
     }
