+++ lib/controllers/mongo.js
@@ -1,10 +1,11 @@
 var xervo = require('../xervo')
 var librarian = require('../common/api').librarian
-var apiUtil = require('@xervo/api-util')
 var userConfig = require('../common/api').userConfig
+var apiUtil = require('../common/api').apiUtil
 
 // -----------------------------------------------------------------------------
 var Mongo = function () {
+
 }
 
 // -----------------------------------------------------------------------------
@@ -25,16 +26,7 @@ Mongo.prototype.create = function (name, creatorId, regionData, callback) {
 
 // -----------------------------------------------------------------------------
 Mongo.prototype.clone = function (db, hosts, callback) {
-  var api = userConfig.data.api_uri.split('://')
-
-  var Client = apiUtil({
-    protocol: api[0],
-    hostname: api[1],
-    version: 'v2',
-    token: userConfig.data.apiKey
-  })
-
-  Client.post('/databases/' + db.id + '/clone', {
+  apiUtil.post('/databases/' + db.id + '/clone', {
     db: db,
     hosts: hosts
   }, callback)
